requirements.txt
Django==4.1.2
djangorestframework==3.14.0
black==22.10.0
isort==5.10.1
